chore: refactor folder structure and updates dependency
