+++ catalyst-gateway/tests/api_tests/api/__init__.py
@@ -1,0 +1,4 @@
+from integration import CAT_GATEWAY_TEST_URL
+
+def cat_gateway_endpoint_url(endpoint: str):
+    return f"{CAT_GATEWAY_TEST_URL}/{endpoint}"
