+++ catalyst-gateway/tests/api_tests/integration/__init__.py
@@ -17,5 +17,3 @@ except KeyError:
 EVENT_DB_TEST_URL = os.environ["EVENT_DB_TEST_URL"]
 CAT_GATEWAY_TEST_URL = os.environ["CAT_GATEWAY_TEST_URL"]
 
-def cat_gateway_endpoint_url(endpoint: str):
-    return f"{CAT_GATEWAY_TEST_URL}/{endpoint}"
