+++ catalyst-gateway/tests/api_tests/utils/auth.py
